# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Isekai",
        "description": "Servicio REST que expone información de un mundo isekai ficticio.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/info/genders": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Obtener listado de géneros",
                "operationId": "305a61324199609b2015d29394aa1f0d",
                "responses": {
                    "200": {
                        "description": "Listado de géneros obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "F"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Femenino"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay información de géneros disponible",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno no manejado",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/info/species": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Obtener listado de especies",
                "operationId": "3dddab0dc5e1c4061893cb8c09972d09",
                "responses": {
                    "200": {
                        "description": "Listado de especies obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "HU"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Humanos"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay información de especies disponible",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno no manejado",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "instance": {
                                            "type": "string",
                                            "example": "https://example.com/"
                                        },
                                        "properties": {
                                            "type": "object",
                                            "example": {
                                                "additionalProp1": "string",
                                                "additionalProp2": "string",
                                                "additionalProp3": "string"
                                            },
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}